Refactor parser into separate file
